Split up tab files and added next and cancel current buttons for admins
